Initial react install + popular movie component + API fetching + styling
Installed react for the first time and it was more simple than I had anticipated coming over from react native Expo which constantly crashed. After finishing installing everything I fetched the movie data, I had already done this exact process from the same website so it was not difficult at all.
What took the most time I would say was the styling, first I went with display grid but it was not feeling right so I switched to flex which immediately improved the look and responsiveness of the website. Another problem I ran into was that the moviePoster was 3 pixels wider than the image, this was very frustrating but I could not waste more time on it so I just set the class to the default image size, might come back to that if I have time. I also need to update the README instead of typing comments like this 😅. But that's for another day!